move WebSocket listeners to Jellyfin.Api
